Started turning trapeze into grab target
Attempted to make a camera follow script, could not decide on a method I was ahppy with so decided against it. Started stting up grab targets on the trapeze, but range detection is not cooperating and is causing a freakout. I believe the fact that the grab target is moving is the likely culprit.